fix issue with quotes in filenames and empty files

+++ extractJPG.py
@@ -14,8 +14,10 @@ if not os.path.isdir(outputDir):
 
 count = 0
 for diskImage in os.listdir(imageDir):
+	#for limiting images for testing purposes
 	count = count + 1
-	if count > 0:
+	if count == 2:
+	#if diskImage == "disk6ba01c70-4900-4012-a393-8cc59c817b25.dd":
 		print diskImage
 		diskOutput = os.path.join(outputDir, diskImage)
 		if not os.path.isdir(diskOutput):
@@ -23,6 +25,29 @@ for diskImage in os.listdir(imageDir):
 		outputXML = os.path.join(workingDir, diskImage + ".xml")
 		fiwalk = Popen(["fiwalk", "-z", "-X", outputXML, os.path.join(imageDir, diskImage)], shell=False, stdout=PIPE, stderr=PIPE)
 		stdout, stderr = fiwalk.communicate()
+
+		fullNamesList = []
+		fullNameSwitch = False
+		"""for joliet extensions ISO9660 only, not udf
+		isoinfo = Popen(["isoinfo", "-f", "-J", "-i", os.path.join(imageDir, diskImage)], shell=False, stdout=PIPE, stderr=PIPE)
+		stdout, stderr = isoinfo.communicate()
+		if len(stderr) > 1:
+			print stderr
+			fullNameSwitch = False
+		if len(stdout) > 1:
+			print stdout
+			fullNameSwitch = True
+			for line in stdout.split("\n"):
+				fullNamesList.append(line.split(";")[0])
+		"""
+
+		"""for debugging
+		fullNameOut = open("fullNameOut.txt", "w")
+		for name in fullNamesList:
+			fullNameOut.write("%s\n" % name)
+		fullNameOut.close()
+		"""
+
 		parser = ET.XMLParser(remove_blank_text=True)
 		fiwalkInput = ET.parse(outputXML, parser)
 		fiwalkXML = fiwalkInput.getroot()
@@ -39,14 +64,21 @@ for diskImage in os.listdir(imageDir):
 								fileList.append(os.path.splitext(os.path.join(os.path.split(os.path.dirname(filename))[0], os.path.basename(filename)).lower())[0])
 				fileCount = len(volume.findall(ns + "fileobject"))
 				if fileCount == 0:
-					print "ERROR: no file owitbject in " + diskImage
+					print "ERROR: no file object in " + diskImage
+					errorLog = open("errorLog.txt", "w")
+					errorLog.write("\nERROR: no file objects found in " + diskImage)
+					errotLog.close()
 				for fileobject in volume:
 					if fileobject.tag == ns + "fileobject":
 						filename = fileobject.find(ns + "filename").text
 						inode = fileobject.find(ns + "inode").text
 						filePath = os.path.dirname(filename)
 						#print filePath
-						newPath = os.path.join(diskOutput, filePath)
+						if fullNameSwitch == True:
+
+							newPath = os.path.join(diskOutput, filePath)
+						else:
+							newPath = os.path.join(diskOutput, filePath)
 						if not os.path.isdir(newPath):
 							os.makedirs(newPath)
 						extention = filename.lower()[-4:]
@@ -55,10 +87,12 @@ for diskImage in os.listdir(imageDir):
 						if filename.lower().endswith(".jpg") or filename.lower().endswith(".jpeg") or filename.lower().endswith(".png"):
 							if int(fileobject.find(ns + "filesize").text) > 2000000:
 								outfile = os.path.join(newPath, os.path.basename(filename))
-								icatCmd = "icat -f iso9660 -i raw " + os.path.join(imageDir, diskImage) + " " + inode + " > " + outfile
+								icatCmd = "icat -f iso9660 -i raw \"" + os.path.join(imageDir, diskImage) + "\" " + inode + " > \"" + outfile + "\""
 								#print icatCmd
 								icat = Popen(icatCmd, shell=True, stdout=PIPE, stderr=PIPE)
 								stdout, stderr = icat.communicate()
+								if len(stderr) > 0:
+									print "icat JPEG error: " + stderr
 						#get all raw images
 						elif extention in extentionList:
 							if os.path.splitext(filename)[0].lower() in fileList:
@@ -67,23 +101,27 @@ for diskImage in os.listdir(imageDir):
 								print "jpg already created for " + os.path.basename(filename)
 							else:
 								outfile = os.path.join(newPath, os.path.basename(filename))
-								icatCmd = "icat -f iso9660 -i raw " + os.path.join(imageDir, diskImage) + " " + inode + " > " + outfile
+								icatCmd = "icat -f iso9660 -i raw \"" + os.path.join(imageDir, diskImage) + "\" " + inode + " > \"" + outfile + "\""
 								#print icatCmd
 								icat = Popen(icatCmd, shell=True, stdout=PIPE, stderr=PIPE)
 								stdout, stderr = icat.communicate()
+								if len(stderr) > 0:
+									print "icat raw error: " + stderr
 				
 				#convert raw images
 				for root, dirs, files in os.walk(newPath):
 					for file in files:
-						fileExt = os.path.splitext(file)[1]
-						if fileExt.lower() in extentionList:
-							print "converting " + file
-							convertCmd = ["convert " + os.path.join(root, file) + " " + os.path.join(root, os.path.splitext(file)[0]) + ".JPG"]
-							convert = Popen(convertCmd, shell=True, stdout=PIPE, stderr=PIPE)
-							stdout, stderr = convert.communicate()
-							os.remove(os.path.join(root, file))
-							#with Image(filename=os.path.join(root, file)) as img:
-								#print img.format
+						if os.path.getsize(os.path.join(root, file)) > 0:
+							fileExt = os.path.splitext(file)[1]
+							if fileExt.lower() in extentionList:
+								print "converting " + file
+								convertCmd = ["convert \"" + os.path.join(root, file) + "\" \"" + os.path.join(root, os.path.splitext(file)[0]) + ".JPG\""]
+								convert = Popen(convertCmd, shell=True, stdout=PIPE, stderr=PIPE)
+								stdout, stderr = convert.communicate()
+								if len(stderr) > 0:
+									print "JPEG error: " + stderr
+								else:
+									os.remove(os.path.join(root, file))
 
 
 		os.remove(outputXML)
